move mmap, poll, fix nits
