+++ c-scape/src/fs/realpath.rs
@@ -24,7 +24,7 @@ unsafe extern "C" fn realpath(path: *const c_char, resolved_path: *mut c_char) -
                     set_errno(Errno(libc::ENOMEM));
                     return null_mut();
                 }
-                std::slice::from_raw_parts_mut(ptr, len).copy_from_slice(&buf[..len]);
+                core::slice::from_raw_parts_mut(ptr, len).copy_from_slice(&buf[..len]);
                 *ptr.add(len) = b'\0';
                 ptr.cast::<c_char>()
             } else {
